- Upgraded MediaWiki to version 1.12.0.   The 1.10.x series isn't listed anymore at mediawiki.org under "Current Versions", so I did a major upgrade to the 1.12.x series. - Revert frik85's changes to "subsys_login.php" in r28600.   They cause the Login system to loop forever, when you use the "Special:Userlogin" page in the Wiki, because $roscms_intern_webserver_roscms was not initialized. We cannot include "custom.php" for this as we don't know the relative path to it.   Another issue of this change was the removal of the if( $subsys == "wiki" ) block, when initializing the Wiki session. This check has to be done, otherwise subsystems like phpBB will run into problems. (their relative path to the Wiki directory might not be "../wiki")   But as reactos.org doesn't show these problems, it looks like this change was never applied to the Web Server.... (and I can be absolutely sure that it causes these problems, because I did a similar mistake before) - Update my documentation about updating the MediaWiki software for the ReactOS Website, I missed a critical part there.
git-svn-id: file:///srv/svn/reactos/trunk@32813 97493ccd-5924-5043-b1f5-66cb403b36ce

+++ web/reactos.org/htdocs/wiki/includes/BagOStuff.php
@@ -73,6 +73,11 @@ class BagOStuff {
 		return true;
 	}
 
+	function keys() {
+		/* stub */
+		return array();
+	}
+
 	/* *** Emulated functions *** */
 	/* Better performance can likely be got with custom written versions */
 	function get_multi($keys) {
@@ -172,7 +177,7 @@ class HashBagOStuff extends BagOStuff {
 	*/
 	var $bag;
 
-	function HashBagOStuff() {
+	function __construct() {
 		$this->bag = array();
 	}
 
@@ -201,6 +206,10 @@ class HashBagOStuff extends BagOStuff {
 			return false;
 		unset($this->bag[$key]);
 		return true;
+	}
+
+	function keys() {
+		return array_keys( $this->bag );
 	}
 }
 
@@ -253,6 +262,9 @@ abstract class SqlBagOStuff extends BagOStuff {
 	}
 
 	function set($key,$value,$exptime=0) {
+		if ( wfReadOnly() ) {
+			return false;
+		}
 		$exptime = intval($exptime);
 		if($exptime < 0) $exptime = 0;
 		if($exptime == 0) {
@@ -272,9 +284,25 @@ abstract class SqlBagOStuff extends BagOStuff {
 	}
 
 	function delete($key,$time=0) {
+		if ( wfReadOnly() ) {
+			return false;
+		}
 		$this->_query(
 			"DELETE FROM $0 WHERE keyname='$1'", $key );
 		return true; /* ? */
+	}
+
+	function keys() {
+		$res = $this->_query( "SELECT keyname FROM $0" );
+		if(!$res) {
+			$this->_debug("keys: ** error: " . $this->_dberror($res) . " **");
+			return array();
+		}
+		$result = array();
+		while( $row = $this->_fetchobject($res) ) {
+			$result[] = $row->keyname;
+		}
+		return $result;
 	}
 
 	function getTableName() {
@@ -339,12 +367,18 @@ abstract class SqlBagOStuff extends BagOStuff {
 
 	function expireall() {
 		/* Remove any items that have expired */
+		if ( wfReadOnly() ) {
+			return false;
+		}
 		$now = $this->_fromunixtime( time() );
 		$this->_query( "DELETE FROM $0 WHERE exptime < '$now'" );
 	}
 
 	function deleteall() {
 		/* Clear *all* items from cache table */
+		if ( wfReadOnly() ) {
+			return false;
+		}
 		$this->_query( "DELETE FROM $0" );
 	}
 
@@ -553,6 +587,52 @@ class eAccelBagOStuff extends BagOStuff {
 }
 
 /**
+ * Wrapper for XCache object caching functions; identical interface
+ * to the APC wrapper
+ */
+class XCacheBagOStuff extends BagOStuff {
+
+	/**
+	 * Get a value from the XCache object cache
+	 *
+	 * @param string $key Cache key
+	 * @return mixed
+	 */
+	public function get( $key ) {
+		$val = xcache_get( $key );
+		if( is_string( $val ) )
+			$val = unserialize( $val );
+		return $val;
+	}
+	
+	/**
+	 * Store a value in the XCache object cache
+	 *
+	 * @param string $key Cache key
+	 * @param mixed $value Object to store
+	 * @param int $expire Expiration time
+	 * @return bool
+	 */
+	public function set( $key, $value, $expire = 0 ) {
+		xcache_set( $key, serialize( $value ), $expire );
+		return true;
+	}
+	
+	/**
+	 * Remove a value from the XCache object cache
+	 *
+	 * @param string $key Cache key
+	 * @param int $time Not used in this implementation
+	 * @return bool
+	 */
+	public function delete( $key, $time = 0 ) {
+		xcache_unset( $key );
+		return true;
+	}
+	
+}
+
+/**
  * @todo document
  */
 class DBABagOStuff extends BagOStuff {
@@ -651,6 +731,7 @@ class DBABagOStuff extends BagOStuff {
 
 	function delete( $key, $time = 0 ) {
 		wfProfileIn( __METHOD__ );
+		wfDebug( __METHOD__."($key)\n" );
 		$handle = $this->getWriter();
 		if ( !$handle ) {
 			return false;
@@ -684,6 +765,19 @@ class DBABagOStuff extends BagOStuff {
 		wfProfileOut( __METHOD__ );
 		return $ret;
 	}
+
+	function keys() {
+		$reader = $this->getReader();
+		$k1 = dba_firstkey( $reader );
+		if( !$k1 ) {
+			return array();
+		}
+		$result[] = $k1;
+		while( $key = dba_nextkey( $reader ) ) {
+			$result[] = $key;
+		}
+		return $result;
+	}
 }
 	
-?>
+
